backend/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
kafka-python
requests